+++ test/engine/engine_core_smooth_test.cc
@@ -252,7 +252,7 @@ TEST_F(CoreSmoothTest, WeldRatioMultipleConstraints) {
 
 // Test Cartesian position control using site transmission with refsite
 TEST_F(CoreSmoothTest, RefsiteBringsToPose) {
-  constexpr char kRefsitePath[] = "engine/testdata/refsite.xml";
+  constexpr char kRefsitePath[] = "engine/testdata/actuation/refsite.xml";
   const std::string xml_path = GetTestDataFilePath(kRefsitePath);
   mjModel* model = mj_loadXML(xml_path.c_str(), nullptr, 0, 0);
   ASSERT_THAT(model, NotNull());
@@ -265,7 +265,7 @@ TEST_F(CoreSmoothTest, RefsiteBringsToPose) {
   mju_copy3(data->ctrl+3, targetrot);
 
   // step for 5 seconds
-  while (data->time < 5) {
+  while (data->time < 10) {
     mj_step(model, data);
   }
 
@@ -273,14 +273,14 @@ TEST_F(CoreSmoothTest, RefsiteBringsToPose) {
   int refsite_id = mj_name2id(model, mjOBJ_SITE, "reference");
   int site_id = mj_name2id(model, mjOBJ_SITE, "end_effector");
 
-  // check that position matches target to within 1e-5 length units
-  double tol_pos = 1e-5;
+  // check that position matches target to within 1e-3 length units
+  double tol_pos = 1e-3;
   mjtNum relpos[3];
   mju_sub3(relpos, data->site_xpos+3*site_id, data->site_xpos+3*refsite_id);
   EXPECT_THAT(relpos, Pointwise(DoubleNear(tol_pos), targetpos));
 
-  // check that orientation matches target to within 1e-3 radians
-  double tol_rot = 1e-3;
+  // check that orientation matches target to within 0.06 radians
+  double tol_rot = 0.06;
   mjtNum site_xquat[4], refsite_xquat[4], relrot[3];
   mju_mat2Quat(refsite_xquat, data->site_xmat+9*refsite_id);
   mju_mat2Quat(site_xquat, data->site_xmat+9*site_id);